refactor: rename board-words.ts to get-board-words.ts, modelOptions to sessionOptions
board-words.ts/getBoardWords was the suggestions/ folder's only file
naming exception to its file-name-equals-main-export convention.
Separately, the hook's local modelOptions() half-collides with the
library's LanguageModelOptions type; the surrounding comments already
call the thing it configures "the session."
